demo: worker-free, edge-portable serving for /todos
The demo's only dynamic route served flash-free HTML via createInlineFlightRenderer,
which owns a long-lived html-worker (worker_threads). That worker exists solely to
render HTML under the opposite React condition — the one thing edge runtimes can't
do — and it pays a thread + React-init cost at boot.

Drop it. /todos document requests now fall through to the prerendered static shell
(what a CDN serves); the client hydrates by fetching the route's live per-request
flight, which is rendered in-process. No worker_threads, no cross-condition HTML
render. The tradeoff: /todos is no longer flash-free on document load (live data
appears just after client render) — which is exactly the portable ceiling.

- rename renderTodosWithInlineFlight.server.tsx -> renderTodosFlight.server.tsx;
  remove renderTodosHtml + createInlineFlightRenderer (and the Html/globalCss only
  it used)
- start.tsx: document request returns null (static shell); only the flight is dynamic
- client.tsx: correct the initial-render comment + dead filename reference

Verified locally against a prod build: clean boot with no worker, /todos document =
static shell, /todos/index.rsc = live flight (reflects a DB insert with no restart),
server actions round-trip through the sealed gate, forged action id rejected.
